+++ include/altacore/ast.hpp
@@ -23,5 +23,6 @@
 #include "ast/import-statement.hpp"
 #include "ast/function-call-expression.hpp"
 #include "ast/string-literal-node.hpp"
+#include "ast/function-declaration-node.hpp"
 
 #endif // ALTACORE_AST_HPP