+++ include/altacore/ast/function-declaration-node.hpp
@@ -1,0 +1,33 @@
+#ifndef ALTACORE_AST_FUNCTION_DECLARATION_NODE_HPP
+#define ALTACORE_AST_FUNCTION_DECLARATION_NODE_HPP
+
+#include "statement-node.hpp"
+#include "block-node.hpp"
+#include "parameter.hpp"
+#include "type.hpp"
+#include "../det/function.hpp"
+#include "../det/scope.hpp"
+#include <vector>
+#include <string>
+
+namespace AltaCore {
+  namespace AST {
+    class FunctionDeclarationNode: public StatementNode {
+      public:
+        virtual const NodeType nodeType();
+
+        std::string name;
+        std::vector<std::shared_ptr<Parameter>> parameters;
+        std::shared_ptr<Type> returnType;
+        std::vector<std::string> modifiers;
+
+        std::shared_ptr<DET::Function> $function;
+
+        FunctionDeclarationNode(std::string name, std::vector<std::shared_ptr<Parameter>> parameters, std::shared_ptr<Type> returnType, std::vector<std::string> modifiers);
+
+        virtual void detail(std::shared_ptr<DET::Scope> scope);
+    };
+  };
+};
+
+#endif // ALTACORE_AST_FUNCTION_DECLARATION_NODE_HPP