+++ CHANGELOG.md
@@ -3,12 +3,15 @@ All the changes for Alta's frontend/core functionality (parser, lexer, AST, DET,
 
 This project follows [semantic versioning](https://semver.org).
 
-## [Unreleased]
+## [0.7.0] - 2018-11-28
 ### Added
 #### DET, AST, Palo (parser)
   * String literal support
-    * Hey hey, we're a hop, skip, and a jump away from `printf`-ing :smile:
-      * All we need to do is add function declaration support
+  * Function declaration support
+### Fixed
+#### Preprocessor
+  * Fixed escaped character support in string literals
+    * Before, we were accessing the wrong vector (`results` instead of `input`)
 
 ## [0.6.0] - 2018-11-27
 ### Added